ok, there, all bad print statements cleaned out...
fixed the problems with logging
part of the problem was I kept typing:

/m perlbot #logging <channel> on

instead of:

/m perlbot #logging <channel> yes

I figured this was normal human behavior, so I added a couple cases to
PerlbotCore.pm to handle that kind of thing.

I updated Chan.pm and Logs.pm to open log files in a more sane way.

Thanks, Ian